requirements.txt
opencv-python
numpy
tqdm
Pillow
torch
torchvision
matplotlib
ultralytics
anvil-uplink
ffmpeg-python
roboflow